refactor: rename all server endpoints to use same file names
