+++ api/server/services/Endpoints/assistants/initialize.spec.js
@@ -2,7 +2,7 @@
 const { HttpsProxyAgent } = require('https-proxy-agent');
 const { ErrorTypes } = require('librechat-data-provider');
 const { getUserKey, getUserKeyExpiry, getUserKeyValues } = require('~/server/services/UserService');
-const initializeClient = require('./initializeClient');
+const initializeClient = require('./initalize');
 // const { OpenAIClient } = require('~/app');
 
 jest.mock('~/server/services/UserService', () => ({
